Reorganize Codebase
- [+] feat(middleware): add custom_next.go file with helper functions
- [+] fix(helper.go): remove redundant helper functions from file
